four of four tests passing
